# Refactor available players in HR