--- a/src/matching/__init__.py
+++ b/src/matching/__init__.py
@@ -3,5 +3,5 @@
 from .matching import Matching
 from .player import Player
 from .game import Game
-from .solvers import StableMarriage, HospitalResident
+from .solvers import HospitalResident, StableMarriage, StudentAllocation
 from .version import __version__