--- a/src/matching/player.py
+++ b/src/matching/player.py
@@ -18,9 +18,8 @@ class Player:
     Attributes
     ==========
     matching : `Player` or `list` of `Player` instances
-        The current match(es) to the player. This is initialised as `None` or an
-        empty list (depending on the type of game) and will remain that way at
-        least until the player's game is solved.
+        The current match(es) to the player. This is initialised as `None` and
+        will remain that way at least until the player is introduced to a game.
     """
 
     def __init__(self, name, pref_names, capacity=1):
@@ -30,8 +29,6 @@ class Player:
         self._pref_names = pref_names
         self.capacity = capacity
         self.matching = None
-        if self.capacity > 1:
-            self.matching = []
 
     def __repr__(self):
 
@@ -40,12 +37,12 @@ class Player:
     def get_favourite_name(self):
         """ Get the player's favourite name. """
 
-        if isinstance(self.matching, list):
+        try:
             for name in self.pref_names:
                 if name not in (match.name for match in self.matching):
                     return name
 
-        if isinstance(self.matching, (Player, type(None))):
+        except TypeError:
             return self.pref_names[0]
 
         return None
@@ -55,43 +52,55 @@ class Player:
 
         fave_name = self.get_favourite_name()
 
-        if not self.matching or isinstance(self.matching, Player):
+        try:
+            for other in others:
+                if other.name == fave_name and other not in self.matching:
+                    return other
+
+        except TypeError:
             for other in others:
                 if other.name == fave_name:
                     return other
-
-        for other in others:
-            if other.name == fave_name and other not in self.matching:
-                return other
 
         return None
 
     def match(self, other):
         """ Assign other to be matched to the player. """
 
-        if isinstance(self.matching, list):
+        try:
             self.matching.append(other)
-        else:
+            self.matching.sort(key=lambda m: self.pref_names.index(m.name))
+
+        except AttributeError:
             self.matching = other
 
     def unmatch(self, other):
         """ Remove other from the player's current matching. """
 
-        if isinstance(self.matching, list):
+        try:
             self.matching.remove(other)
-        else:
+
+        except AttributeError:
             self.matching = None
 
     def get_worst_match_idx(self):
         """ Get the preference list index of the player's worst current match.
         """
 
-        if isinstance(self.matching, list):
-            return max(
-                (self.pref_names.index(other.name) for other in self.matching)
-            )
+        try:
+            return self.pref_names.index(self.matching[-1].name)
 
-        return self.pref_names.index(self.matching.name)
+        except TypeError:
+            return self.pref_names.index(self.matching.name)
+
+    def get_worst_match(self):
+        """ Get the worst current match to a player. """
+
+        try:
+            return self.matching[-1]
+
+        except TypeError:
+            return self.matching
 
     def forget(self, other):
         """ Forget another player by removing their name from the player's
@@ -99,10 +108,8 @@ class Player:
 
         self.pref_names.remove(other.name)
 
-    def get_successors(self, others, idx=None):
-        """ Get all the successors either to the current match (for simple
-        games) or the worst current match (for capacitated games). This match is
-        at position `idx` in the player's preference list. """
+    def get_successors(self, others):
+        """ Get all the successors to the worst current match of a player. """
 
         idx = self.get_worst_match_idx()
         return [