--- a/src/matching/solvers/__init__.py
+++ b/src/matching/solvers/__init__.py
@@ -2,3 +2,4 @@
 
 from .hospital_resident import HospitalResident, hospital_resident
 from .stable_marriage import StableMarriage, stable_marriage
+from .student_allocation import StudentAllocation