--- a/tests/student_allocation/test_solver.py
+++ b/tests/student_allocation/test_solver.py
@@ -6,12 +6,8 @@ from .params import STUDENT_ALLOCATION, _make_game
 
 
 @STUDENT_ALLOCATION
-def test_init(
-    student_names,
-    proj_lect_dict,
-    project_capacities,
-    lecturer_capacities,
-    seed,
+def _test_init(
+    student_names, proj_lect_dict, project_capacities, lecturer_capacities, seed
 ):
     """ Test that an instance of StudentAllocation is created correctly. """