--- a/tests/unit/test_matching.py
+++ b/tests/unit/test_matching.py
@@ -36,6 +36,29 @@ def test_repr():
     matching = Matching()
     assert repr(matching) == "{}"
 
+    matching = Matching(dictionary)
+    assert repr(matching) == str(dictionary)
+
+
+def test_keys():
+    """ Check a Matching can have its `keys` accessed. """
+
+    matching = Matching()
+    assert list(matching.keys()) == []
+
+    matching = Matching(dictionary)
+    assert list(matching.keys()) == suitors
+
+
+def test_values():
+    """ Check a Matching can have its `values` accessed. """
+
+    matching = Matching()
+    assert list(matching.values()) == []
+
+    matching = Matching(dictionary)
+    assert list(matching.values()) == reviewers
+
 
 def test_getitem():
     """ Check that you can access items in a Matching correctly. """