Following #26, some tweaking has been included to speed up the processes in solving an instance of HR.

The main change is that instead of finding all the "free" players at the end of each iteration, they are added and dropped on the fly during the matching, unmatching and forgetting stages of the algorithm(s).

The effect of these tweaks was tested in a similar fashion to that done in #26 for a game with 400 residents and 20 hospitals. The runtime on my machine decreased from ~8s to ~2.5s for resident-optimality and from ~10s to ~0.5s for hospital-optimality.